+++ CHANGELOG.md
@@ -2,15 +2,21 @@
 
 ## [0.7.0] - TODO
 ### Fixes
+- Fixed that `ContainerNetwork`s were using the `name` for hostnames instead of the `host_name` that
+  was meant for that purpose
 - `Command` stdout copiers no longer panic on invalid utf-8
 
 ### Changes
 - `Command` and `CommandResult` stdout and stderr are now `Vec<u8>` instead of `String`
+- `ContainerNetwork` now adds on a UUID suffix to docker names and hostnames in order to allow
+  running them in parallel
 - `assert_success` is no longer `track_caller`
 - Many dependency updates, use `postcard` internally instead of `bincode`
 
 ### Additions
 - Added `CommandResult::stdout_as_utf8` and some other related functions for convenience
+- Added `ContainerNetwork::terminate_containers` which just terminates containers and not the
+  network
 
 ## [0.6.0] - 2023-10-07
 ### Changes